split test in reworked file
